added comments, a remove and addOrder in accountservice, and added more to the orderView jsp- orderIncrement

+++ VSS-web/src/main/java/com/team33/controllers/AccountController.java
@@ -52,7 +52,7 @@ public class AccountController extends AbstractController{
  
     @RequestMapping("/delete/{accountId}")
     public String removeAccount(@PathVariable("accountId")
-    Long accountId) {
+    Integer accountId) {
  
         accountService.removeAccount(accountId);
