+++ VSS-web/src/main/java/com/team33/entities/dao/AccountDao.java
@@ -15,5 +15,5 @@ public interface AccountDao {
           
 	  public void saveAccount(Account account) throws DataAccessException;
           
-          public void removeAccount(Long accountID);
+          public void removeAccount(Integer accountID);
 }
